backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
databases
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
ccxt==4.2.1
web3==6.15.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
aiofiles==23.1.0
pytest==7.4.0
httpx==0.25.1
gunicorn==20.1.0
sqlalchemy-utils==0.41.1
celery==5.3.0
redis==5.0.0
